+++ src/metatrain/cli/eval.py
@@ -4,7 +4,7 @@ import itertools
 import logging
 import time
 from pathlib import Path
-from typing import Dict, Optional, Union
+from typing import Dict, List, Optional, Union
 
 import numpy as np
 import torch
@@ -24,6 +24,7 @@ from metatrain.utils.data import (
     read_systems,
     unpack_batch,
 )
+from metatrain.utils.data.readers import read_extra_data
 from metatrain.utils.data.target_info import DEPRECATED_METATOMIC_OUTPUT_NAMES
 from metatrain.utils.data.writers import (
     DiskDatasetWriter,
@@ -43,6 +44,7 @@ from metatrain.utils.neighbor_lists import (
 from metatrain.utils.omegaconf import expand_dataset_config
 from metatrain.utils.per_atom import average_by_num_atoms
 from metatrain.utils.pydantic import validate_eval_options
+from metatrain.utils.system_data import get_system_data_transform
 from metatrain.utils.transfer import batch_to
 
 
@@ -114,6 +116,19 @@ def _add_eval_model_parser(subparser: argparse._SubParsersAction) -> None:
         action="store_true",
         help="whether to run consistency checks (default: %(default)s)",
     )
+    parser.add_argument(
+        "--warm-up",
+        action=argparse.BooleanOptionalAction,
+        dest="warm_up",
+        default=True,
+        help=(
+            "whether to do a warm-up of the model before evaluation "
+            "(default: %(default)s). The warm-up is done by running 10 "
+            "model evaluations. This will give timings that are more representative "
+            "of the model's performance on a large dataset. However, it will delay "
+            "evaluation, which might be undesirable for expensive evaluations."
+        ),
+    )
 
 
 def _prepare_eval_model_args(args: argparse.Namespace) -> None:
@@ -136,6 +151,7 @@ def _eval_targets(
     batch_size: int = 1,
     check_consistency: bool = False,
     writer: Optional[Writer] = None,
+    warm_up: bool = True,
 ) -> None:
     """
     Evaluate `model` on `dataset`, accumulate RMSE/MAE, and (if `writer` is provided)
@@ -147,6 +163,7 @@ def _eval_targets(
     :param batch_size: Batch size for evaluation.
     :param check_consistency: Whether to run consistency checks during model evaluation.
     :param writer: Optional writer to write out per-sample predictions.
+    :param warm_up: Whether to do a warm-up of the model before evaluation.
     """
     # Disable static fusion. Besides the fact that atomistic batches have variable
     # sizes, statically fused CUDA kernels cannot allocate new tensors at runtime,
@@ -180,28 +197,40 @@ def _eval_targets(
     # Create a dataloader
     target_keys = list(model.capabilities().outputs.keys())
     requested_neighbor_lists = get_requested_neighbor_lists(model)
-    collate_fn = CollateFn(
-        target_keys,
-        callables=[get_system_with_neighbor_lists_transform(requested_neighbor_lists)],
-    )
+    callables = [get_system_with_neighbor_lists_transform(requested_neighbor_lists)]
+
+    # Attach additional per-system inputs
+    requested_inputs = [
+        name
+        for name, output in model.requested_inputs(use_new_names=True).items()
+        if output.sample_kind == "system"
+    ]
+    if requested_inputs:
+        callables.append(get_system_data_transform(requested_inputs))
+    collate_fn = CollateFn(target_keys, callables=callables)
     dataloader = torch.utils.data.DataLoader(
         dataset, batch_size=batch_size, collate_fn=collate_fn, shuffle=False
     )
+
     rmse_acc = RMSEAccumulator()
     mae_acc = MAEAccumulator()
 
     # Warm-up
-    cycled = itertools.cycle(dataloader)
-    for _ in range(10):
-        batch = unpack_batch(next(cycled))
-        systems = [system.to(device=device, dtype=dtype) for system in batch[0]]
-        evaluate_model(
-            model,
-            systems,
-            options,
-            is_training=False,
-            check_consistency=check_consistency,
-        )
+    if warm_up:
+        logging.info("Warming up the model with 10 batches...")
+        cycled = itertools.cycle(dataloader)
+        for _ in range(10):
+            batch = unpack_batch(next(cycled))
+            systems = [system.to(device=device, dtype=dtype) for system in batch[0]]
+            evaluate_model(
+                model,
+                systems,
+                options,
+                is_training=False,
+                check_consistency=check_consistency,
+            )
+    else:
+        logging.info("Skipping warm-up of the model.")
 
     total_time = 0.0
     timings_per_atom = []
@@ -275,6 +304,7 @@ def eval_model(
     batch_size: int = 1,
     check_consistency: bool = False,
     append: Optional[bool] = None,
+    warm_up: bool = True,
 ) -> None:
     """
     Evaluate an exported model on a given data set.
@@ -289,6 +319,7 @@ def eval_model(
     :param batch_size: Batch size for evaluation.
     :param check_consistency: Whether to run consistency checks during model evaluation.
     :param append: If ``True``, open the output file in append mode.
+    :param warm_up: Whether to do a warm-up of the model before evaluation.
     """
     logging.info("Setting up evaluation set.")
     output = Path(output) if isinstance(output, str) else output
@@ -308,11 +339,6 @@ def eval_model(
         # build the dataset & target-info
         if hasattr(options, "targets"):
             eval_dataset, eval_info_dict, _ = get_dataset(options)
-            eval_systems = (
-                [d.system for d in eval_dataset]
-                if not isinstance(writer, DiskDatasetWriter)
-                else None
-            )
         else:
             if isinstance(writer, DiskDatasetWriter):
                 raise ValueError(
@@ -325,7 +351,7 @@ def eval_model(
             )
 
             # FIXME: this works only for energy models
-            eval_targets: Dict[str, TensorMap] = {}
+            eval_targets: Dict[str, List[TensorMap]] = {}
             eval_info_dict = copy.deepcopy(
                 {
                     name: model_output
@@ -337,7 +363,13 @@ def eval_model(
                 if "energy" in name:
                     model_output.sample_kind = "system"  # type: ignore
 
-            eval_dataset = Dataset.from_dict({"system": eval_systems, **eval_targets})
+            extra_data: Dict[str, List[TensorMap]] = {}
+            if "extra_data" in options:
+                extra_data, _ = read_extra_data(conf=options["extra_data"])
+
+            eval_dataset = Dataset.from_dict(
+                {"system": eval_systems, **eval_targets, **extra_data}
+            )
 
         if "indices" in options:
             eval_indices = options["indices"]
@@ -356,6 +388,7 @@ def eval_model(
                 batch_size=batch_size,
                 check_consistency=check_consistency,
                 writer=writer,
+                warm_up=warm_up,
             )
         except Exception as e:
             raise ArchitectureError(e)